requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2

transformers>=4.11.0
torch>=1.9.0
nltk>=3.6.3
bertopic>=0.9.4
textstat>=0.7.2
shap>=0.39.0
bert-score>=0.3.11

openai>=1.0.0
google-generativeai>=0.1.0

gutenberg-cleaner==0.1.1
requests==2.31.0

python-dotenv>=0.19.0
tqdm>=4.62.3


